+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - fix "Update FieldRoles" does not map some of the field roles in NG format to 7.3 which is used by the Web Components
  - Cart Page
   - fix array to string conversion error in recommendation template 
+ - Configuration
+  - fix Test Connection and Test Upload Connection do not work on a store scope
  
 ## [v4.0.0-rc.3] - 2022.05.13
 ### Add