+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## Unreleased
+## [v3.5.2] - 2022.08.08
 ### Add
 - Product view, Export Preview
   - added button on product page which open new page with preview of the exported values 
@@ -476,6 +476,7 @@
 ### Added
 - Feed Export: Export feed file is now available via separate link
 
+[v3.5.2]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v3.5.2
 [v3.5.1]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v3.5.1
 [v3.5.0]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v3.5.0
 [v3.4.1]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v3.4.1